refactor: move pricing types from lib to wire crate
