+++ lib/src/pricing.rs
@@ -2,37 +2,14 @@ use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
 use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
+use lrzcc_wire::pricing::{
+    FlavorPrice, FlavorPriceCreateData, FlavorPriceInitialize,
+    FlavorPriceModifyData,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorPrice {
-    pub id: u32,
-    pub flavor: u32,
-    pub flavor_name: String,
-    pub user_class: u32,
-    pub unit_price: f64,
-    pub start_time: DateTime<FixedOffset>,
-}
-
-impl Display for FlavorPrice {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!(
-            "FlavorPrice(id={}, flavor={})",
-            self.id, self.flavor_name
-        ))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorPriceInitialize {
-    pub new_flavor_price_count: u32,
-}
 
 pub struct FlavorPriceApi {
     pub url: String,
@@ -63,28 +40,6 @@ impl FlavorPriceListRequest {
             SerializableNone!(),
             StatusCode::OK,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct FlavorPriceCreateData {
-    flavor: u32,
-    // TODO use an enum for this
-    user_class: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    price: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    start_time: Option<DateTime<FixedOffset>>,
-}
-
-impl FlavorPriceCreateData {
-    fn new(flavor: u32, user_class: u32) -> Self {
-        Self {
-            flavor,
-            user_class,
-            price: None,
-            start_time: None,
-        }
     }
 }
 
@@ -130,32 +85,6 @@ impl FlavorPriceCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct FlavorPriceModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    flavor: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    user_class: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    unit_price: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    start_time: Option<DateTime<FixedOffset>>,
-}
-
-impl FlavorPriceModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            flavor: None,
-            user_class: None,
-            unit_price: None,
-            start_time: None,
-        }
     }
 }
